// src/generated/models/Swag_swagitemsModel.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */
export const Swag_swagitemsstatecode = {
  0: 'Active',
  1: 'Inactive'
} as const;
export type Swag_swagitemsstatecode = keyof typeof Swag_swagitemsstatecode;
export const Swag_swagitemsstatuscode = {
  1: 'Active',
  2: 'Inactive'
} as const;
export type Swag_swagitemsstatuscode = keyof typeof Swag_swagitemsstatuscode;
export const Swag_swagitemsswag_sizecategory = {
  100000000: 'S',
  100000001: 'M',
  100000002: 'L'
} as const;
export type Swag_swagitemsswag_sizecategory = keyof typeof Swag_swagitemsswag_sizecategory;

export interface Swag_swagitemsBase {
  importsequencenumber?: string;
  overriddencreatedon?: string;
  statecode: Swag_swagitemsstatecode;
  statuscode?: Swag_swagitemsstatuscode;
  swag_name: string;
  swag_quantityreserved: string;
  swag_quantitytotal: string;
  swag_sizecategory: Swag_swagitemsswag_sizecategory;
  swag_swagitemid: string;
  swag_tiereligibility?: string;
  swag_unitcost: string;
  timezoneruleversionnumber?: string;
  "TransactionCurrencyId@odata.bind"?: string;
  utcconversiontimezonecode?: string;
}

export interface Swag_swagitems extends Swag_swagitemsBase {
  createdbyname?: string;
  createdbyyominame: string;
  createdon?: string;
  createdonbehalfbyname?: string;
  createdonbehalfbyyominame: string;
  exchangerate?: string;
  modifiedbyname?: string;
  modifiedbyyominame: string;
  modifiedon?: string;
  modifiedonbehalfbyname?: string;
  modifiedonbehalfbyyominame: string;
  organizationidname: string;
  statecodename?: string;
  statuscodename?: string;
  swag_sizecategoryname?: string;
  swag_unitcost_base?: string;
  transactioncurrencyidname?: string;
  versionnumber?: string;
  createdby?: object;
  _createdby_value?: string;
  createdonbehalfby?: object;
  _createdonbehalfby_value?: string;
  modifiedby?: object;
  _modifiedby_value?: string;
  modifiedonbehalfby?: object;
  _modifiedonbehalfby_value?: string;
  organizationid?: object;
  _organizationid_value?: string;
  transactioncurrencyid?: object;
  _transactioncurrencyid_value?: string;
}