// src/generated/models/Swag_eventsModel.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */
export const Swag_eventsstatecode = {
  0: 'Active',
  1: 'Inactive'
} as const;
export type Swag_eventsstatecode = keyof typeof Swag_eventsstatecode;
export const Swag_eventsstatuscode = {
  1: 'Active',
  2: 'Inactive'
} as const;
export type Swag_eventsstatuscode = keyof typeof Swag_eventsstatuscode;
export const Swag_eventsswag_inperson = {
  0: 'No',
  1: 'Yes'
} as const;
export type Swag_eventsswag_inperson = keyof typeof Swag_eventsswag_inperson;
export const Swag_eventsswag_status = {
  100000000: 'draft',
  100000001: 'confirmed',
  100000002: 'amended',
  100000003: 'ordered',
  100000004: 'cancelled'
} as const;
export type Swag_eventsswag_status = keyof typeof Swag_eventsswag_status;

export interface Swag_eventsBase {
  importsequencenumber?: string;
  overriddencreatedon?: string;
  statecode: Swag_eventsstatecode;
  statuscode?: Swag_eventsstatuscode;
  swag_attendeecount: string;
  swag_champcount: string;
  swag_eventdate: string;
  swag_eventid: string;
  swag_execcount: string;
  swag_inperson: Swag_eventsswag_inperson;
  swag_sellername: string;
  swag_selleruserid?: string;
  swag_status: Swag_eventsswag_status;
  swag_tier3count: string;
  swag_tier4count: string;
  swag_totalcost?: string;
  swag_tpid?: string;
  timezoneruleversionnumber?: string;
  "TransactionCurrencyId@odata.bind"?: string;
  utcconversiontimezonecode?: string;
}

export interface Swag_events extends Swag_eventsBase {
  createdbyname?: string;
  createdbyyominame: string;
  createdon?: string;
  createdonbehalfbyname?: string;
  createdonbehalfbyyominame: string;
  exchangerate?: string;
  modifiedbyname?: string;
  modifiedbyyominame: string;
  modifiedon?: string;
  modifiedonbehalfbyname?: string;
  modifiedonbehalfbyyominame: string;
  organizationidname: string;
  statecodename?: string;
  statuscodename?: string;
  swag_inpersonname?: string;
  swag_statusname?: string;
  swag_totalcost_base?: string;
  transactioncurrencyidname?: string;
  versionnumber?: string;
  createdby?: object;
  _createdby_value?: string;
  createdonbehalfby?: object;
  _createdonbehalfby_value?: string;
  modifiedby?: object;
  _modifiedby_value?: string;
  modifiedonbehalfby?: object;
  _modifiedonbehalfby_value?: string;
  organizationid?: object;
  _organizationid_value?: string;
  transactioncurrencyid?: object;
  _transactioncurrencyid_value?: string;
}