// src/generated/models/Swag_allocationsModel.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */
export const Swag_allocationsstatecode = {
  0: 'Active',
  1: 'Inactive'
} as const;
export type Swag_allocationsstatecode = keyof typeof Swag_allocationsstatecode;
export const Swag_allocationsstatuscode = {
  1: 'Active',
  2: 'Inactive'
} as const;
export type Swag_allocationsstatuscode = keyof typeof Swag_allocationsstatuscode;
export const Swag_allocationsswag_status = {
  100000000: 'proposed',
  100000001: 'confirmed',
  100000002: 'ordered',
  100000003: 'cancelled'
} as const;
export type Swag_allocationsswag_status = keyof typeof Swag_allocationsswag_status;

export interface Swag_allocationsBase {
  importsequencenumber?: string;
  overriddencreatedon?: string;
  statecode: Swag_allocationsstatecode;
  statuscode?: Swag_allocationsstatuscode;
  swag_allocationid: string;
  "swag_EventId@odata.bind": string;
  "swag_ItemId@odata.bind": string;
  swag_itemname: string;
  swag_quantity: string;
  swag_status?: Swag_allocationsswag_status;
  swag_totalcost?: string;
  timezoneruleversionnumber?: string;
  "TransactionCurrencyId@odata.bind"?: string;
  utcconversiontimezonecode?: string;
}

export interface Swag_allocations extends Swag_allocationsBase {
  createdbyname?: string;
  createdbyyominame: string;
  createdon?: string;
  createdonbehalfbyname?: string;
  createdonbehalfbyyominame: string;
  exchangerate?: string;
  modifiedbyname?: string;
  modifiedbyyominame: string;
  modifiedon?: string;
  modifiedonbehalfbyname?: string;
  modifiedonbehalfbyyominame: string;
  organizationidname: string;
  statecodename?: string;
  statuscodename?: string;
  swag_eventidname?: string;
  swag_itemidname?: string;
  swag_statusname?: string;
  swag_totalcost_base?: string;
  transactioncurrencyidname?: string;
  versionnumber?: string;
  createdby?: object;
  _createdby_value?: string;
  createdonbehalfby?: object;
  _createdonbehalfby_value?: string;
  modifiedby?: object;
  _modifiedby_value?: string;
  modifiedonbehalfby?: object;
  _modifiedonbehalfby_value?: string;
  organizationid?: object;
  _organizationid_value?: string;
  swag_eventid?: object;
  _swag_eventid_value?: string;
  swag_itemid?: object;
  _swag_itemid_value?: string;
  transactioncurrencyid?: object;
  _transactioncurrencyid_value?: string;
}