// src/generated/models/Swag_allocationrulesModel.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */
export const Swag_allocationrulesstatecode = {
  0: 'Active',
  1: 'Inactive'
} as const;
export type Swag_allocationrulesstatecode = keyof typeof Swag_allocationrulesstatecode;
export const Swag_allocationrulesstatuscode = {
  1: 'Active',
  2: 'Inactive'
} as const;
export type Swag_allocationrulesstatuscode = keyof typeof Swag_allocationrulesstatuscode;
export const Swag_allocationrulesswag_eventtype = {
  100000000: 'in_person',
  100000001: 'virtual'
} as const;
export type Swag_allocationrulesswag_eventtype = keyof typeof Swag_allocationrulesswag_eventtype;

export interface Swag_allocationrulesBase {
  importsequencenumber?: string;
  overriddencreatedon?: string;
  statecode: Swag_allocationrulesstatecode;
  statuscode?: Swag_allocationrulesstatuscode;
  swag_allocationruleid: string;
  swag_eventtype: Swag_allocationrulesswag_eventtype;
  swag_rulesjson?: string;
  swag_tiername: string;
  timezoneruleversionnumber?: string;
  utcconversiontimezonecode?: string;
}

export interface Swag_allocationrules extends Swag_allocationrulesBase {
  createdbyname?: string;
  createdbyyominame: string;
  createdon?: string;
  createdonbehalfbyname?: string;
  createdonbehalfbyyominame: string;
  modifiedbyname?: string;
  modifiedbyyominame: string;
  modifiedon?: string;
  modifiedonbehalfbyname?: string;
  modifiedonbehalfbyyominame: string;
  organizationidname: string;
  statecodename?: string;
  statuscodename?: string;
  swag_eventtypename?: string;
  versionnumber?: string;
  createdby?: object;
  _createdby_value?: string;
  createdonbehalfby?: object;
  _createdonbehalfby_value?: string;
  modifiedby?: object;
  _modifiedby_value?: string;
  modifiedonbehalfby?: object;
  _modifiedonbehalfby_value?: string;
  organizationid?: object;
  _organizationid_value?: string;
}