// src/generated/models/Swag_budgettrackersModel.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */
export const Swag_budgettrackersstatecode = {
  0: 'Active',
  1: 'Inactive'
} as const;
export type Swag_budgettrackersstatecode = keyof typeof Swag_budgettrackersstatecode;
export const Swag_budgettrackersstatuscode = {
  1: 'Active',
  2: 'Inactive'
} as const;
export type Swag_budgettrackersstatuscode = keyof typeof Swag_budgettrackersstatuscode;

export interface Swag_budgettrackersBase {
  importsequencenumber?: string;
  overriddencreatedon?: string;
  statecode: Swag_budgettrackersstatecode;
  statuscode?: Swag_budgettrackersstatuscode;
  swag_budgettrackerid: string;
  swag_programname: string;
  swag_programtotal: string;
  swag_quarterend: string;
  swag_quarterstart: string;
  swag_totalreserved: string;
  swag_totalspent: string;
  timezoneruleversionnumber?: string;
  "TransactionCurrencyId@odata.bind"?: string;
  utcconversiontimezonecode?: string;
}

export interface Swag_budgettrackers extends Swag_budgettrackersBase {
  createdbyname?: string;
  createdbyyominame: string;
  createdon?: string;
  createdonbehalfbyname?: string;
  createdonbehalfbyyominame: string;
  exchangerate?: string;
  modifiedbyname?: string;
  modifiedbyyominame: string;
  modifiedon?: string;
  modifiedonbehalfbyname?: string;
  modifiedonbehalfbyyominame: string;
  organizationidname: string;
  statecodename?: string;
  statuscodename?: string;
  swag_programtotal_base?: string;
  swag_totalreserved_base?: string;
  swag_totalspent_base?: string;
  transactioncurrencyidname?: string;
  versionnumber?: string;
  createdby?: object;
  _createdby_value?: string;
  createdonbehalfby?: object;
  _createdonbehalfby_value?: string;
  modifiedby?: object;
  _modifiedby_value?: string;
  modifiedonbehalfby?: object;
  _modifiedonbehalfby_value?: string;
  organizationid?: object;
  _organizationid_value?: string;
  transactioncurrencyid?: object;
  _transactioncurrencyid_value?: string;
}